Accessing the Data Cube

Accessing the Data Cube

Open a data cube

Before one can read data from a cube, it has to be opened. To open a data cube, which is usually accessible through the file system, use the Cube(base_dir) constructor:

using ESDL
c = Cube("/path/to/cube/")

ESDL data cube at /path/to/cube/
Spatial resolution:  4320x2160 at 0.083333 degrees.
Temporal resolution: 2001-01-01T00:00:00 to 2012-01-01T00:00:00 at 8daily time steps
Variables:           aerosol_optical_thickness_1610 aerosol_optical_thickness_550 aerosol_optical_thickness_555 aerosol_optical_thickness_659 aerosol_optical_thickness_865 air_temperature_2m bare_soil_evaporation black_sky_albedo burnt_area c_emissions country_mask evaporation evaporative_stress fractional_snow_cover gross_primary_productivity interception_loss land_surface_temperature latent_energy net_ecosystem_exchange open_water_evaporation ozone potential_evaporation precipitation root_moisture sensible_heat snow_sublimation snow_water_equivalent soil_moisture surface_moisture terrestrial_ecosystem_respiration transpiration water_mask water_vapour white_sky_albedo

This returns a Cube object and prints some basic information (location, resolutions & variable names) about the cube. Using Cube() without any arguments will attempt to access a data cube at the path specified in the environmental variable ENV["ESDL_CUBEDIR"] if applicable or else try to open a remote data cube.

ESDL.CubeAPI.Cube — Type.

Represents a data cube accessible though the file system. The default constructor is

Cube(base_dir)

where base_dir is the data cube's base directory.

Fields

base_dir the cube parent directory

config the cube's static configuration CubeConfig

dataset_files a list of datasets in the cube

var_name_to_var_index basically the inverse of dataset_files

source

ESDL.CubeAPI.Cube — Method.

Cube(base_dir::AbstractString)

Represents a data cube accessible though the file system.

Argument

base_dir the parent directory of the data cube to be opened.

source

ESDL.CubeAPI.Cube — Method.

Cube(;resolution="low")

Empty constructor with a default keyword argument for the resolution. Attempts to access default data cube locations in the file system. If unsuccessful attempts to open a RemoteCube.

Keyword argument

resolution Can be "low" or "high". Defaults to "low".

source

Getting a data handle

The next step after opening the data cube, is to select the extents of the data cube to be further subject to processing. The following code example serves to illustrate this step:

var = ["c_emissions", "air_temperature_2m"]
time = (Date("2001-01-01"), Date("2001-12-31"))
lons = (30,32), lats = (50,51)
cubedata = getCubeData(c, longitude=lons, latitude=lats, time=time, variable=var)

Data Cube view with the following dimensions
Lon                 Axis with 8 Elements from 30.125 to 31.875
Lat                 Axis with 4 Elements from 50.875 to 50.125
Time                Axis with 46 Elements from 2001-01-01 to 2001-12-27
Variable            Axis with 2 elements: c_emissions air_temperature_2m
Total size: 14.38 KB

This returns a view of the data cube, on which further calculations can be applied. The extents here are all data for the variables "cemissions" and "airtemperature_2m" between 50° and 51° North and 30° and 32° East for the year 2001. For southern latitudes and western longitudes use negative values. All keyword arguments default to the full range, so calling getCubeData() without keyword arguments will return a view of the whole data cube.

Here you can start to do calculations on your sub-cube. See either Analysis for a list of methods provided by this framework or Applying custom functions to apply your own functions on the cube. If you just want to visualize the cube see the section Plotting.

Note that the data wasn't read yet. In case you want to load some data into memory and store it in a Julia array, just use square-bracket indexing. For example, to read the first time step (2001-01-01) of the second variable (air_temperature_2m) as a Lon-Lat array, do

cubedata[:,:,1,2]

8×4 reshape(::Array{Union{Missing, Float32},4}, 8, 4) with eltype Union{Missing, Float32}:
 274.301  274.409  274.458  274.499
 274.299  274.411  274.473  274.529
 274.297  274.412  274.488  274.558
 274.296  274.414  274.503  274.588
 274.294  274.415  274.518  274.618
 274.28   274.412  274.533  274.649
 274.266  274.409  274.547  274.681
 274.251  274.406  274.562  274.713

ESDL.CubeAPI.getCubeData — Function.

getCubeData(cube::Cube; variable, time, latitude, longitude)

Returns a SubCube object which represents a view of the original data cube. The following keyword arguments are accepted:

variable a variable index, variable name string or an iterable returning multiple of these, like: (var1, var2, ...)

time a single Date object or a 2-element iterable of Date objects, like: (start_time, end_time)

latitude a single latitude value or a 2-element iterable, like: (first_latitude, last_latitude)

longitude a single longitude value or a 2-element iterable, like (first_longitude, last_longitude)

region specify a country or SREX region by name or ISO_A3 code. Overrides latitude and longitude for AOI selection. Type ?ESDL.known_regions to see a list of pre-defined areas

source

Opening Remote Data Cubes

If you just want to try the ESDL and don't have access to the full data set, you can open a remote cube through a THREDDS server. All you need is a working internet connection to do this:

ESDL.CubeAPI.RemoteCube — Method.

RemoteCube(;resolution="low", url)

Keyword arguments

resolution Can be "low" or "high". Defaults to "low".

url A custom URL for the remote connection can be specified here. Defaults to the official remote server.

source

This will open the remote cube and calling getCubeData will return a cube view that you can process.

Important: Please be nice to our servers. Use this only for testing the ESDL software for a limited amount of data (reading maps at single time steps) or time series in lon-lat boxes of size 1° x 1°.

Cube Types

While the getCubeData command returns an object of type SubCube, which represents a view into the ESDC, other cube operations will return different types of data cubes. The returned type will depend on the size of the returned cube. If it is small enough to fit into memory, it will be a CubeMem, otherwise a MmapCube. All these types of data cubes share the same interface defined by AbstractCubeData, which means you can index them, do calculation using mapCube or plot them using the commands described in Plotting.

ESDL.Cubes.AbstractCubeData — Type.

AbstractCubeData{T,N}

Supertype of all cubes. T is the data type of the cube and N the number of dimensions. Beware that an AbstractCubeData does not implement the AbstractArray interface. However, the ESDL functions mapCube, readCubeData, plotMAP and plotXY will work on any subtype of AbstractCubeData

source

ESDL.Cubes.CubeMem — Type.

CubeMem{T,N} <: AbstractCubeMem{T,N}

An in-memory data cube. It is returned by applying mapCube when the output cube is small enough to fit in memory or by explicitly calling readCubeData on any type of cube.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

source

ESDL.CubeAPI.SubCube — Type.

immutable SubCube{T,C} <: AbstractCubeData{T,4}

A view into the data cube of a single variable. Is the type returned by the mapCube function.

Fields

cube::C Parent cube

variable selected variable

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis LonAxis object

latAxis LatAxis object

timeAxis TimeAxis object

source

ESDL.CubeAPI.SubCubeV — Type.

immutable SubCubeV{T, C} <: AbstractCubeData{T,4}

A view into the data cube with multiple variables. Returned by the mapCube function.

Fields

cube::C Parent cube

variable list of selected variables

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis LonAxis object

latAxis LatAxis object

timeAxis TimeAxis object

varAxis VariableAxis object

source

ESDL.Cubes.MmapCube — Type.

MmapCube{T,N}

Defines a memory-mapped data cube which is stored on disk. Is generally returned by mapCube applications.

source

Cube Dimensions

Dimensions are an essential part of each Cube in ESDL. Every dimension that a cube has, is associated with an axis that stores the values of the dimension. For example, a LatitudeAxis will contain a field values representing the chosen latitudes. Similarly, a VariableAxis will contain a list of Variable names. Axes types are divided in CategoricalAxis and RangeAxis types. All of them are subtypes of the abstract type CubeAxis.

ESDL.Cubes.Axes.CubeAxis — Type.

abstract CubeAxis{T} <: AbstractCubeData{T,1}

Supertype of all axes. Every CubeAxis is 1D Cube itself and can be passed to mapCube operationes. Although all cube axes are instances of the parametric typealias CategoricalAxis and RangeAxis, there are some typealiases defined to provide shorter and more convenient names for commonly used cube axes. Here is a list of the aliases:

Categorical Axes

VariableAxis represents different variables

SpatialPointAxis represents a list of coordinates

CountryAxis countries

ScaleAxis time scales after time series decomposition

QuantileAxis represents different quantiles

Continuous Axes

LonAxis longitudes

LatAxis latitudes

TimeAxis time

MSCAxis time step inside a year (for seasonal statistics)

source

ESDL.Cubes.Axes.CategoricalAxis — Type.

CategoricalAxis{T,S}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol). The default constructor is:

CategoricalAxis(axname::String,values::Vector{T})

source

ESDL.Cubes.Axes.RangeAxis — Type.

RangeAxis{T,S,R}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol) and R the type of the range which is used to represent the axis values. The default constructor is:

RangeAxis(axname::String,values::Range{T})

source

List of known regions

ESDL.CubeAPI.known_regions — Constant.

List of Continents

Africa

Asia

Australia

Europe

North America

South America

List of SREX regions

Short Name	Long Name

ALA	Alaska/N.W. Canada

AMZ	Amazon

CAM	Central America/Mexico

CAR	small islands regions Caribbean

CAS	Central Asia

CEU	Central Europe

CGI	Canada/Greenland/Iceland

CNA	Central North America

EAF	East Africa

EAS	East Asia

ENA	East North America

MED	South Europe/Mediterranean

NAS	North Asia

NAU	North Australia

NEB	North-East Brazil

NEU	North Europe

SAF	Southern Africa

SAH	Sahara

SAS	South Asia

SAU	South Australia/New Zealand

SEA	Southeast Asia

SSA	Southeastern South America

TIB	Tibetan Plateau

WAF	West Africa

WAS	West Asia

WNA	West North America

WSA	West Coast South America

ANT	Antarctica

ARC	Arctic

NTP	Pacific Islands region

STP	Southern Topical Pacific

ETP	Pacific Islands region

WIO	West Indian Ocean

List of countries

Short Name	Long Name

ABW	Aruba

AFG	Afghanistan

AGO	Angola

AIA	Anguilla

ALB	Albania

ALA	Aland

AND	Andorra

ARE	United Arab Emirates

ARG	Argentina

ARM	Armenia

ASM	American Samoa

ATA	Antarctica

Ashm	Ashmore and Cartier Is.

ATF	Fr. S. and Antarctic Lands

ATG	Antigua and Barb.

AUS	Australia

AUT	Austria

AZE	Azerbaijan

BDI	Burundi

BEL	Belgium

BEN	Benin

BFA	Burkina Faso

BGD	Bangladesh

BGR	Bulgaria

BHR	Bahrain

BHS	Bahamas

BIH	Bosnia and Herz.

BLM	St. Barthelemy

BLR	Belarus

BLZ	Belize

BMU	Bermuda

BOL	Bolivia

BRA	Brazil

BRB	Barbados

BRN	Brunei

BTN	Bhutan

BWA	Botswana

CAF	Central African Rep.

CAN	Canada

CHE	Switzerland

CHL	Chile

CHN	China

CIV	Ivory Coast

CLP	Clipperton I.

CMR	Cameroon

CNM

COD	Congo (Kinshasa)

COG	Congo (Brazzaville)

COK	Cook Is.

COL	Colombia

COM	Comoros

CPV	Cape Verde

CRI	Costa Rica

CSI	Coral Sea Is.

CUB	Cuba

CUW	Curacao

CYM	Cayman Is.

CYN	N. Cyprus

CYP	Cyprus

CZE	Czech Rep.

DEU	Germany

DJI	Djibouti

DMA	Dominica

DNK	Denmark

DOM	Dominican Rep.

DZA	Algeria

ECU	Ecuador

EGY	Egypt

ERI	Eritrea

ESB	Dhekelia

ESP	Spain

EST	Estonia

ETH	Ethiopia

FIN	Finland

FJI	Fiji

FLK	Falkland Is.

FRA	France

FRO	Faroe Is.

FSM	Micronesia

GAB	Gabon

Gaza	Gaza

GBR	United Kingdom

GEO	Georgia

GGY	Guernsey

GHA	Ghana

GIB	Gibraltar

GIN	Guinea

GMB	Gambia

GNB	Guinea Bissau

GNQ	Eq. Guinea

GRC	Greece

GRD	Grenada

GRL	Greenland

GTM	Guatemala

GUM	Guam

GUY	Guyana

HKG	Hong Kong

HMD	Heard I. and McDonald Is.

HND	Honduras

HRV	Croatia

HTI	Haiti

HUN	Hungary

IDN	Indonesia

IMN	Isle of Man

IND	India

IOA	Indian Ocean Ter.

IOT	Br. Indian Ocean Ter.

IRL	Ireland

IRN	Iran

IRQ	Iraq

ISL	Iceland

ISR	Israel

ITA	Italy

JAM	Jamaica

JEY	Jersey

JOR	Jordan

JPN	Japan

KAB	Baykonur

KAS

KAZ	Kazakhstan

KEN	Kenya

KGZ	Kyrgyzstan

KHM	Cambodia

KIR	Kiribati

KNA	St. Kitts and Nevis

KNM

KOR	S. Korea

KOS	Kosovo

KWT	Kuwait

LAO	Laos

LBN	Lebanon

LBR	Liberia

LBY	Libya

LCA	Saint Lucia

LIE	Liechtenstein

LKA	Sri Lanka

LSO	Lesotho

LTU	Lithuania

LUX	Luxembourg

LVA	Latvia

MAC	Macau

MAF	St. Martin

MAR	Morocco

MCO	Monaco

MDA	Moldova

MDG	Madagascar

MDV	Maldives

MEX	Mexico

MHL	Marshall Is.

MKD	Macedonia

MLI	Mali

MLT	Malta

MMR	Myanmar

MNE	Montenegro

MNG	Mongolia

MNP	N. Mariana Is.

MOZ	Mozambique

MRT	Mauritania

MSR	Montserrat

MUS	Mauritius

MWI	Malawi

MYS	Malaysia

NAM	Namibia

NCL	New Caledonia

NER	Niger

NFK	Norfolk Island

NGA	Nigeria

NIC	Nicaragua

NIU	Niue

NLD	Netherlands

NOR	Norway

NPL	Nepal

NRU	Nauru

NZL	New Zealand

OMN	Oman

PAK	Pakistan

PAN	Panama

PCN	Pitcairn Is.

PER	Peru

PHL	Philippines

PLW	Palau

PNG	Papua New Guinea

POL	Poland

PRI	Puerto Rico

PRK	N. Korea

PRT	Portugal

PRY	Paraguay

PYF	Fr. Polynesia

QAT	Qatar

ROU	Romania

RUS	Russia

RWA	Rwanda

SAH	W. Sahara

SAU	Saudi Arabia

SDN	Sudan

SDS	S. Sudan

SEN	Senegal

SGP	Singapore

SGS	S. Geo. and S. Sandw. Is.

SHN	Saint Helena

SLB	Solomon Is.

SLE	Sierra Leone

SLV	El Salvador

SMR	San Marino

SOL	Somaliland

SOM	Somalia

SPM	St. Pierre and Miquelon

SRB	Serbia

STP	Sao Tome and Principe

SUR	Suriname

SVK	Slovakia

SVN	Slovenia

SWE	Sweden

SWZ	Swaziland

SXM	Sint Maarten

SYC	Seychelles

SYR	Syria

TCA	Turks and Caicos Is.

TCD	Chad

TGO	Togo

THA	Thailand

TJK	Tajikistan

TKM	Turkmenistan

TLS	East Timor

TON	Tonga

TTO	Trinidad and Tobago

TUN	Tunisia

TUR	Turkey

TUV	Tuvalu

TWN	Taiwan

TZA	Tanzania

UGA	Uganda

UKR	Ukraine

UMI	U.S. Minor Outlying Is.

URY	Uruguay

USA	United States

USG	Guantanamo Bay USNB

UZB	Uzbekistan

VAT	Vatican

VCT	St. Vin. and Gren.

VEN	Venezuela

VGB	British Virgin Is.

VIR	U.S. Virgin Is.

VNM	Vietnam

VUT	Vanuatu

PSE	West Bank

WLF	Wallis and Futuna

WSB	Akrotiri

WSM	Samoa

YEM	Yemen

ZAF	South Africa

ZMB	Zambia

ZWE	Zimbabwe

source