CubeAPI

CubeAPI

ESDL.CubeAPI.known_regions

ESDL.CubeAPI.Cube

ESDL.CubeAPI.CubeConfig

ESDL.CubeAPI.RemoteCube

ESDL.CubeAPI.SubCube

ESDL.CubeAPI.SubCubePerm

ESDL.CubeAPI.SubCubeStatic

ESDL.CubeAPI.SubCubeStaticPerm

ESDL.CubeAPI.SubCubeV

ESDL.CubeAPI.SubCubeVPerm

ESDL.CubeAPI.getCubeData

ESDL.CubeAPI.getCubeData

ESDL.CubeAPI.getTimeRanges

ESDL.CubeAPI.getTimesToRead

ESDL.CubeAPI.showVarInfo

ESDL.CubeAPI.known_regions — Constant.

List of Continents

Africa

Asia

Australia

Europe

North America

South America

List of SREX regions

Short Name	Long Name

ALA	Alaska/N.W. Canada

AMZ	Amazon

CAM	Central America/Mexico

CAR	small islands regions Caribbean

CAS	Central Asia

CEU	Central Europe

CGI	Canada/Greenland/Iceland

CNA	Central North America

EAF	East Africa

EAS	East Asia

ENA	East North America

MED	South Europe/Mediterranean

NAS	North Asia

NAU	North Australia

NEB	North-East Brazil

NEU	North Europe

SAF	Southern Africa

SAH	Sahara

SAS	South Asia

SAU	South Australia/New Zealand

SEA	Southeast Asia

SSA	Southeastern South America

TIB	Tibetan Plateau

WAF	West Africa

WAS	West Asia

WNA	West North America

WSA	West Coast South America

ANT	Antarctica

ARC	Arctic

NTP	Pacific Islands region

STP	Southern Topical Pacific

ETP	Pacific Islands region

WIO	West Indian Ocean

List of countries

Short Name	Long Name

ABW	Aruba

AFG	Afghanistan

AGO	Angola

AIA	Anguilla

ALB	Albania

ALA	Aland

AND	Andorra

ARE	United Arab Emirates

ARG	Argentina

ARM	Armenia

ASM	American Samoa

ATA	Antarctica

Ashm	Ashmore and Cartier Is.

ATF	Fr. S. and Antarctic Lands

ATG	Antigua and Barb.

AUS	Australia

AUT	Austria

AZE	Azerbaijan

BDI	Burundi

BEL	Belgium

BEN	Benin

BFA	Burkina Faso

BGD	Bangladesh

BGR	Bulgaria

BHR	Bahrain

BHS	Bahamas

BIH	Bosnia and Herz.

BLM	St. Barthelemy

BLR	Belarus

BLZ	Belize

BMU	Bermuda

BOL	Bolivia

BRA	Brazil

BRB	Barbados

BRN	Brunei

BTN	Bhutan

BWA	Botswana

CAF	Central African Rep.

CAN	Canada

CHE	Switzerland

CHL	Chile

CHN	China

CIV	Ivory Coast

CLP	Clipperton I.

CMR	Cameroon

CNM

COD	Congo (Kinshasa)

COG	Congo (Brazzaville)

COK	Cook Is.

COL	Colombia

COM	Comoros

CPV	Cape Verde

CRI	Costa Rica

CSI	Coral Sea Is.

CUB	Cuba

CUW	Curacao

CYM	Cayman Is.

CYN	N. Cyprus

CYP	Cyprus

CZE	Czech Rep.

DEU	Germany

DJI	Djibouti

DMA	Dominica

DNK	Denmark

DOM	Dominican Rep.

DZA	Algeria

ECU	Ecuador

EGY	Egypt

ERI	Eritrea

ESB	Dhekelia

ESP	Spain

EST	Estonia

ETH	Ethiopia

FIN	Finland

FJI	Fiji

FLK	Falkland Is.

FRA	France

FRO	Faroe Is.

FSM	Micronesia

GAB	Gabon

Gaza	Gaza

GBR	United Kingdom

GEO	Georgia

GGY	Guernsey

GHA	Ghana

GIB	Gibraltar

GIN	Guinea

GMB	Gambia

GNB	Guinea Bissau

GNQ	Eq. Guinea

GRC	Greece

GRD	Grenada

GRL	Greenland

GTM	Guatemala

GUM	Guam

GUY	Guyana

HKG	Hong Kong

HMD	Heard I. and McDonald Is.

HND	Honduras

HRV	Croatia

HTI	Haiti

HUN	Hungary

IDN	Indonesia

IMN	Isle of Man

IND	India

IOA	Indian Ocean Ter.

IOT	Br. Indian Ocean Ter.

IRL	Ireland

IRN	Iran

IRQ	Iraq

ISL	Iceland

ISR	Israel

ITA	Italy

JAM	Jamaica

JEY	Jersey

JOR	Jordan

JPN	Japan

KAB	Baykonur

KAS

KAZ	Kazakhstan

KEN	Kenya

KGZ	Kyrgyzstan

KHM	Cambodia

KIR	Kiribati

KNA	St. Kitts and Nevis

KNM

KOR	S. Korea

KOS	Kosovo

KWT	Kuwait

LAO	Laos

LBN	Lebanon

LBR	Liberia

LBY	Libya

LCA	Saint Lucia

LIE	Liechtenstein

LKA	Sri Lanka

LSO	Lesotho

LTU	Lithuania

LUX	Luxembourg

LVA	Latvia

MAC	Macau

MAF	St. Martin

MAR	Morocco

MCO	Monaco

MDA	Moldova

MDG	Madagascar

MDV	Maldives

MEX	Mexico

MHL	Marshall Is.

MKD	Macedonia

MLI	Mali

MLT	Malta

MMR	Myanmar

MNE	Montenegro

MNG	Mongolia

MNP	N. Mariana Is.

MOZ	Mozambique

MRT	Mauritania

MSR	Montserrat

MUS	Mauritius

MWI	Malawi

MYS	Malaysia

NAM	Namibia

NCL	New Caledonia

NER	Niger

NFK	Norfolk Island

NGA	Nigeria

NIC	Nicaragua

NIU	Niue

NLD	Netherlands

NOR	Norway

NPL	Nepal

NRU	Nauru

NZL	New Zealand

OMN	Oman

PAK	Pakistan

PAN	Panama

PCN	Pitcairn Is.

PER	Peru

PHL	Philippines

PLW	Palau

PNG	Papua New Guinea

POL	Poland

PRI	Puerto Rico

PRK	N. Korea

PRT	Portugal

PRY	Paraguay

PYF	Fr. Polynesia

QAT	Qatar

ROU	Romania

RUS	Russia

RWA	Rwanda

SAH	W. Sahara

SAU	Saudi Arabia

SDN	Sudan

SDS	S. Sudan

SEN	Senegal

SGP	Singapore

SGS	S. Geo. and S. Sandw. Is.

SHN	Saint Helena

SLB	Solomon Is.

SLE	Sierra Leone

SLV	El Salvador

SMR	San Marino

SOL	Somaliland

SOM	Somalia

SPM	St. Pierre and Miquelon

SRB	Serbia

STP	Sao Tome and Principe

SUR	Suriname

SVK	Slovakia

SVN	Slovenia

SWE	Sweden

SWZ	Swaziland

SXM	Sint Maarten

SYC	Seychelles

SYR	Syria

TCA	Turks and Caicos Is.

TCD	Chad

TGO	Togo

THA	Thailand

TJK	Tajikistan

TKM	Turkmenistan

TLS	East Timor

TON	Tonga

TTO	Trinidad and Tobago

TUN	Tunisia

TUR	Turkey

TUV	Tuvalu

TWN	Taiwan

TZA	Tanzania

UGA	Uganda

UKR	Ukraine

UMI	U.S. Minor Outlying Is.

URY	Uruguay

USA	United States

USG	Guantanamo Bay USNB

UZB	Uzbekistan

VAT	Vatican

VCT	St. Vin. and Gren.

VEN	Venezuela

VGB	British Virgin Is.

VIR	U.S. Virgin Is.

VNM	Vietnam

VUT	Vanuatu

PSE	West Bank

WLF	Wallis and Futuna

WSB	Akrotiri

WSM	Samoa

YEM	Yemen

ZAF	South Africa

ZMB	Zambia

ZWE	Zimbabwe

source

ESDL.CubeAPI.Cube — Type.

Represents a data cube accessible though the file system. The default constructor is

Cube(base_dir)

where base_dir is the datacube's base directory.

Fields

base_dir the cube parent directory

config the cube's static configuration CubeConfig

dataset_files a list of datasets in the cube

var_name_to_var_index basically the inverse of dataset_files

source

ESDL.CubeAPI.RemoteCube — Type.

Represents a remote data cube accessible through THREDDS. The default constructor is

RemoteCube(base_url)

where base_url is the datacube's base url.

Fields

base_url the cube parent directory

var_name_to_var_index basically the inverse of dataset_files

dataset_files a list of datasets in the cube

dataset_paths a list of urls pointing to the different data sets

config the cube's static configuration CubeConfig

using ESDL
ds=remoteCube()

source

ESDL.CubeAPI.getCubeData — Method.

getCubeData(cube::Cube; variable, time, latitude, longitude)

Returns a SubCube object which represents a view of the original data cube. The following keyword arguments are accepted:

variable a variable index, variable name string or an iterable returning multiple of these, like: (var1, var2, ...)

time a single Date object or a 2-element iterable of Date objects, like: (start_time, end_time)

latitude a single latitude value or a 2-element iterable, like: (first_latitude, last_latitude)

longitude a single longitude value or a 2-element iterable, like (first_longitude, last_longitude)

region specify a country or SREX region by name or ISO_A3 code. Overrides latitude and longitude for AOI selection. Type ?ESDL.known_regions to see a list of pre-defined areas

source

ESDL.CubeAPI.getTimeRanges — Method.

Returns a vector of Date objects giving the time indices returned by a respective call to getCubeData.

source

ESDL.CubeAPI.showVarInfo — Method.

showVarInfo(cube)

Shows the metadata and citation information on variables contained in a cube.

source

ESDL.CubeAPI.CubeConfig — Type.

A data cube's static configuration information.

spatial_res: The spatial image resolution in degree.

grid_x0: The fixed grid X offset (longitude direction).

grid_y0: The fixed grid Y offset (latitude direction).

grid_width: The fixed grid width in pixels (longitude direction).

grid_height: The fixed grid height in pixels (latitude direction).

static_data:

temporal_res: The temporal resolution in days.

ref_time: A Date value which defines the units in which time values are given, namely days since ref_time.

start_time: The start time of the first image of any variable in the cube given as Date value.

$None$ means unlimited.

end_time: The end time of the last image of any variable in the cube given as Date value.

$None$ means unlimited.

variables: A list of variable names to be included in the cube.

file_format: The file format used. Must be one of 'NETCDF4', 'NETCDF4CLASSIC', 'NETCDF3CLASSIC'

or 'NETCDF3_64BIT'.

compression: Whether the data should be compressed.

source

ESDL.CubeAPI.SubCube — Type.

immutable SubCube{T,C} <: AbstractCubeData{T,4}

A view into the data cube of a single variable. Is the type returned by the mapCube function.

Fields

cube::C Parent cube

variable selected variable

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis LonAxis object

latAxis LatAxis object

timeAxis TimeAxis object

source

ESDL.CubeAPI.SubCubePerm — Type.

immutable SubCubePerm{T} <: AbstractCubeData{T,3}

Representation of a SubCube with perumted dimensions

Fields

parent Parent SubCube

perm the permutation

iperm the inverse permutation

source

ESDL.CubeAPI.SubCubeStatic — Type.

immutable SubCubeStatic{T, C} <: AbstractCubeData{T,2}

A view into the data cube with a single static variable. Returned by the mapCube function.

Fields

cube::C Parent cube

variable list of selected variables

sub_grid representation of the subgrid indices

time representation of the selected time steps

lonAxis

latAxis

source

ESDL.CubeAPI.SubCubeStaticPerm — Type.

immutable SubCubeStaticPerm{T} <: AbstractCubeData{T,2}

Representation of a SubCubeStatic with permuted dimensions.

source

ESDL.CubeAPI.SubCubeV — Type.

immutable SubCubeV{T, C} <: AbstractCubeData{T,4}

A view into the data cube with multiple variables. Returned by the mapCube function.

Fields

cube::C Parent cube

variable list of selected variables

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis LonAxis object

latAxis LatAxis object

timeAxis TimeAxis object

varAxis VariableAxis object

source

ESDL.CubeAPI.SubCubeVPerm — Type.

immutable SubCubeVPerm{T} <: AbstractCubeData{T,4}

Representation of a SubCubeV with permuted dimensions.

source

ESDL.CubeAPI.getTimesToRead — Method.

Function to get the years and times to read from user input.

source

ESDL.Cubes._read — Method.

Add a function to read some CubeData in a permuted way, we will make a copy here for simplicity, however, this might change in the future

source