Cubes

Cubes

ESDL.Cubes.AbstractCubeData

ESDL.Cubes.AbstractCubeMem

ESDL.Cubes.AbstractSubCube

ESDL.Cubes.ConcatCube

ESDL.Cubes.CubeMem

ESDL.Cubes.MmapCube

ESDL.Cubes._read

ESDL.Cubes.caxes

ESDL.Cubes.chunkoffset

ESDL.Cubes.concatenateCubes

ESDL.Cubes.cubechunks

ESDL.Cubes.getSingVal

ESDL.Cubes.gethandle

ESDL.Cubes.loadCube

ESDL.Cubes.loadCube

ESDL.Cubes.readCubeData

ESDL.Cubes.rmCube

ESDL.Cubes.saveCube

ESDL.Cubes.saveCube

ESDL.Cubes.saveCube

ESDL.Cubes.subsetCubeData

ESDL.Cubes — Module.

The functions provided by ESDL are supposed to work on different types of cubes. This module defines the interface for all Data types that

source

ESDL.Cubes.AbstractCubeData — Type.

AbstractCubeData{T,N}

Supertype of all cubes. T is the data type of the cube and N the number of dimensions. Beware that an AbstractCubeData does not implement the AbstractArray interface. However, the ESDL functions ESDL.DAT.mapCube, reduceCube, readCubeData, plotMAP and plotXY will work on any subtype of AbstractCubeData

source

ESDL.Cubes.AbstractCubeMem — Type.

Supertype of all in-memory representations of a data cube

source

ESDL.Cubes.AbstractSubCube — Type.

Supertype of all subtypes of the original data cube

source

ESDL.Cubes.ConcatCube — Type.

ConcatCube

Concatenate a list of cubes of the same type and axes to a composite cube.

source

ESDL.Cubes.CubeMem — Type.

CubeMem{T,N} <: AbstractCubeMem{T,N}

An in-memory data cube. It is returned by applying mapCube when the output cube is small enough to fit in memory or by explicitly calling readCubeData on any type of cube.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

source

ESDL.Cubes.MmapCube — Type.

MmapCube{T,N}

Defines a Memory-Mapped data cube which is stored on disk. Is generally returned by mapCube applications.

source

ESDL.Cubes.caxes — Method.

Returns the axes of a Cube

source

ESDL.Cubes.concatenateCubes — Method.

function concatenateCubes(cubelist, catAxis::CategoricalAxis)

Concatenates a vector of datacubes that have identical axes to a new single cube along the new axis catAxis

source

ESDL.Cubes.getSingVal — Method.

getSingVal reads a single point from the cube's data

source

ESDL.Cubes.loadCube — Method.

loadCube(name::String)

Loads a cube that was previously saved with saveCube. Returns a TempCube object.

source

ESDL.Cubes.readCubeData — Method.

readCubeData(cube::AbstractCubeData)

source

ESDL.Cubes.rmCube — Method.

rmCube(name::String)

Deletes a memory-mapped data cube.

source

ESDL.Cubes.saveCube — Method.

saveCube(c::AbstractCubeData, name::String)

Permanently saves a data cube to disk by either moving the folder out of the tmp directory (for TempCubes) or by storing the data to disk (for CubeMems)

source

ESDL.Cubes.saveCube — Method.

saveCube(cube,name::String)

Save a MmapCube or CubeMem to the folder name in the ESDL working directory.

See also loadCube, ESDLdir

source

Base.ndims — Method.

Number of dimensions

source

ESDL.Cubes.chunkoffset — Method.

Offset of the first chunk

source

ESDL.Cubes.cubechunks — Method.

Chunks, if given

source

ESDL.Cubes.gethandle — Method.

gethandle(c::AbstractCubeData, [block_size])

Returns an indexable handle to the data.

source

ESDL.Cubes.subsetCubeData — Function.

This function calculates a subset of a cube's data

source