DAT

DAT

ESDL.DAT.DATConfig

ESDL.DAT.InDims

ESDL.DAT.InnerObj

ESDL.DAT.InputCube

ESDL.DAT.OutDims

ESDL.DAT.OutputCube

ESDL.DAT.cmpcachmisses

ESDL.DAT.fittable

ESDL.DAT.getFrontPerm

ESDL.DAT.getLoopCacheSize

ESDL.DAT.mapCube

ESDL.DAT.mapCube

ESDL.DAT.@CubeTable

ESDL.DAT.InputCube — Type.

Internal representation of an input cube for DAT operations

source

ESDL.DAT.OutputCube — Type.

Internal representation of an output cube for DAT operations

source

ESDL.DAT.fittable — Method.

fittable(tab,o,fitsym;by=(),weight=nothing)

Loops through an iterable table tab and thereby fitting an OnlineStat o with the values specified through fitsym. Optionally one can specify a field (or tuple) to group by. Any groupby specifier can either be a symbol denoting the entry to group by or an anynymous function calculating the group from a table row.

For example the following would caluclate a weighted mean over a cube weighted by grid cell area and grouped by country and month

fittable(iter,WeightedMean,:tair,weight=(i->cosd(i.lat)),by=(i->month(i.time),:country))

source

ESDL.DAT.mapCube — Method.

mapCube(fun, cube, addargs...;kwargs)

Map a given function fun over slices of the data cube cube.

Keyword arguments

max_cache=1e7 maximum size of blocks that are read into memory, defaults to approx 10Mb

outtype::DataType output data type of the operation

indims::InDims List of input cube descriptors of type InDims for each input data cube

outdims::OutDims List of output cube descriptors of type OutDims for each output cube

inplace does the function write to an output array inplace or return a single value> defaults to true

ispar boolean to determine if parallelisation should be applied, defaults to true if workers are available.

outfolder a folder where the output cube is stroed, defaults to the result of ESDLdir()

showprog boolean indicating if a ProgressMeter shall be shown ESDL.Cubes.AbstractCubeData

kwargs additional keyword arguments passed to the inner function

The first argument is always the function to be applied, the second is the input cube or a tuple input cubes if needed.

source

ESDL.DAT.@CubeTable — Macro.

@CubeTable input_vars...

Macro to turn a DataCube object into an iterable table. Takes a list of as arguments, specified either by a cube variable name alone or by a name=cube expression. For example @CubeTable cube1 country=cube2 would generate a Table with the entries cube1 and country, where cube1 contains the values of cube1 and country the values of cube2. The cubes are matched and broadcasted along their axes like in mapCube.

In addition, one can specify axes=(ax1,ax2...) when one wants to include the values of certain xes in the table. For example the command @CubeTable tair=cube1 axes=(lon,lat,time) would produce an iterator over a data structure with entries tair, lon, lat and time.

Lastly there is an option to specify which axis shall be the fastest changing when iterating over the cube. For example @CubeTable cube1 fastest=time will ensure that the iterator will always loop over consecutive time steps of the same location.

source

ESDL.DAT.DATConfig — Type.

Configuration object of a DAT process. This holds all necessary information to perform the calculations It contains the following fields:

incubes::Vector{AbstractCubeData} The input data cubes

outcube::AbstractCubeData The output data cube

indims::Vector{Tuple} Tuples of input axis types

outdims::Tuple Tuple of output axis types

axlists::Vector{Vector{CubeAxis}} Axes of the input data cubes

inAxes::Vector{Vector{CubeAxis}}

outAxes::Vector{CubeAxis}

LoopAxes::Vector{CubeAxis}

axlistOut::Vector{CubeAxis}

ispar::Bool

isMem::Vector{Bool}

inCubesH

outCubeH

source

ESDL.DAT.InnerObj — Type.

Object to pass to InnerLoop, this condenses the most important information about the calculation into a type so that specific code can be generated by the @generated function

source

ESDL.DAT.cmpcachmisses — Method.

Function that compares two cache miss specifiers by their importance

source

ESDL.DAT.getFrontPerm — Method.

Calculate an axis permutation that brings the wanted dimensions to the front

source

ESDL.DAT.getLoopCacheSize — Method.

Calculate optimal Cache size to DAT operation

source