Analysis

Apply functions

List of functions

Seasonal cycles

CABLAB.Proc.MSC.getMSC — Function.

Calculate the mean seasonal cycle of xin and write the output to xout.

source

CABLAB.Proc.MSC.getMedSC — Method.

Calculate the median seasonal cycle of xin and write the output to xout.

source

CABLAB.Proc.MSC.removeMSC — Method.

removeMSC

Removes the mean annual cycle from each time series.

Call signature

mapCube(removeMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis

source

Outlier detection

CABLAB.Proc.Outlier.DAT_detectAnomalies! — Method.

DAT_detectAnomalies!

A simple wrapper around the function detectAnomalies! from the MultivariateAnomalies package.

Call signature

mapCube(DAT_detectAnomalies!, cube, methods, trainArray)

cube data cube with a axes: TimeAxis, VariableAxis

methods vector of methods to be applied, choose from: KDE,T2,REC,KNN-Gamma,KNN-Delta,SVDD,KNFST

trainArray a matrix of nsample x nvar, to estimate the training parameters for the model. Ideally does not contain any extreme values

Input Axes TimeAxis, Variableaxis

Output Axes TimeAxis, Methodaxis

For an example on how to apply this function, see this notebook.

source

Simple Statistics

Time series decomposition

Cube transformations