Analysis

The CABLAB package comes with a list of predefined methods for statistical analysis. The functions are defined to work on specific axes, for example a function that removes the mean annual cycle will alway work an the time axis. It does not matter which other axes are defined in the input cube, the function will simply loop over these. All the functions are called using the mapCube function.

CABLAB.DAT.mapCube — Function.

mapCube(fun, cube, addargs...;kwargs)

Map a given function fun over slices of the data cube cube.

Keyword arguments

max_cache=1e7 maximum size of blocks that are read into memory, defaults to approx 10Mb

outfolder folder to write output to if a TempCube is created, defaults to joinpath(CABLABdir(),"tmp",SomeRandomName)

outtype::DataType output data type of the operation

indims::Tuple{Tuple{Vararg{CubeAxis}}} List of input axis types for each input data cube

outdims::Tuple List of output axes, can be either an axis type that has a default constructor or an instance of a CubeAxis

inmissing::Tuple How to treat missing values in input data for each input cube. Possible values are :nullable :mask :nan or a value that is inserted for missing data, defaults to :mask

outmissing How are missing values written to the output array, possible values are :nullable, :mask, :nan, defaults to :mask

no_ocean should values containing ocean data be omitted

inplace does the function write to an output array inplace or return a single value> defaults to true

kwargs additional keyword arguments passed to the inner function

The first argument is always the function to be applied, the second is the input cube or a tuple input cubes if needed. If the function to be applied is registered (either as part of CABLAB or through registerDATFunction), all of the keyword arguments have reasonable defaults and don't need to be supplied. Some of the function still need additional arguments or keyword arguments as is stated in the documentation.

If you want to call mapCube directly on an unregistered function, please have a look at Applying custom functions to get an idea about the usage of the input and output dimensions etc.

source

The function will then be applied to the whole cube in a memory-efficient way, which means that chunks of data are read, processed and then saved in the output cube. Whether the output cube is a TempCube or a CubeMem is decided by the system, depending on if the calculation is parallel, and how large the output cube is.

Here follows a list of analysis function included in this package. If you have implemented or wrapped a method that might be of interest to a broader community, please feel free to open a pull request.

Seasonal cycles

CABLAB.Proc.MSC.gapFillMSC — Method.

gapFillMSC

Fills missing values of each time series with the mean annual cycle.

Call signature

mapCube(gapFillMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis

source

CABLAB.Proc.MSC.getMSC — Function.

getMSC

Returns the mean annual cycle from each time series.

Call signature

mapCube(getMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes MSCaxis

source

CABLAB.Proc.MSC.getMedSC — Method.

getMedMSC

Returns the median annual cycle from each time series.

Call signature

mapCube(getMedMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes MSCaxis

source

CABLAB.Proc.MSC.removeMSC — Method.

removeMSC

Removes the mean annual cycle from each time series.

Call signature

mapCube(removeMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis

source

Outlier detection

CABLAB.Proc.Outlier.DAT_detectAnomalies! — Method.

DAT_detectAnomalies!

A simple wrapper around the function detectAnomalies! from the MultivariateAnomalies package.

Call signature

mapCube(DAT_detectAnomalies!, cube, methods, trainArray)

cube data cube with a axes: TimeAxis, VariableAxis

methods vector of methods to be applied, choose from: KDE,T2,REC,KNN-Gamma,KNN-Delta,SVDD,KNFST

trainArray a matrix of nsample x nvar, to estimate the training parameters for the model. Ideally does not contain any extreme values

Input Axes TimeAxis, Variableaxis

Output Axes TimeAxis, Methodaxis

For an example on how to apply this function, see this notebook.

source

Simple Statistics

Another typcial use case is the application of basic statistics like sum, mean and std. We provide a convenience function reduceCube

CABLAB.DAT.reduceCube — Function.

reduceCube(f::Function, cube, dim::Type{T<:CubeAxis};kwargs...)

Apply a reduction function f on slices of the cube cube. The dimension(s) are specified through dim, which is either an Axis type or a tuple of axis types. Keyword arguments are passed to mapCube or, if unknown passed again to f. It is assumed that f takes an array input and returns a single value.

source

Additional simple statistics functions are:

CABLAB.Proc.Stats.normalizeTS — Method.

normalizeTS

Normalize a time series to zeros mean and unit variance

Call signature

mapCube(normalizeTS, cube)

cube data cube with a axes: TimeAxis

Input Axes TimeAxis

Output Axes TimeAxis

source

CABLAB.Proc.Stats.timespacequantiles — Method.

timespacequantiles

Calculate quantiles from a space time data cube. This is usually called on a subset of data returned by sampleLandPoints.

Call signature

mapCube(timespacequantiles, cube, quantiles)

cube data cube with a axes: TimeAxis, SpatialPointAxis

quantiles a vector of quantile values to calculate

Input Axes TimeAxis, SpatialPointAxis

Output Axes QuantileAxis

Calculating exact quantiles from data that don't fit into memory is quite a problem. One solution we provide here is to simply subsample your data and then get the quantiles from a smaller dataset.

For an example on how to apply this function, see this notebook.

source

Time series decomposition

CABLAB.Proc.TSDecomposition.filterTSFFT — Method.

filterTSFFT

Filter each time series using a Fourier filter and return the decomposed series in 4 time windows (Trend, Long-Term Variability, Annual Cycle, Fast Oscillations)

Call signature

mapCube(filterTSFFT, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis, TimeScaleaxis

source

Cube transformations