Analysis

The CABLAB package comes with a list of predefined methods for statistical analysis. The functions are defined to work on specific axes, for example a function that removes the mean annual cycle will alway work an the time axis. It does not matter which other axes are defined in the input cube, the function will simply loop over these. All the functions are called using the mapCube function.

CABLAB.DAT.mapCube — Function.

mapCube(fun, cube, addargs...;kwargs)

Map a given function fun over slices of the data cube cube.

Keyword arguments

max_cache=1e7 maximum size of blocks that are read into memory, defaults to approx 10Mb

outtype::DataType output data type of the operation

indims::Tuple{Tuple{Vararg{CubeAxis}}} List of input axis types for each input data cube

outdims::Tuple List of output axes, can be either an axis type that has a default constructor or an instance of a CubeAxis

inmissing::Tuple How to treat missing values in input data for each input cube. Possible values are :nullable :mask :nan or a value that is inserted for missing data, defaults to :mask

outmissing How are missing values written to the output array, possible values are :nullable, :mask, :nan, defaults to :mask

no_ocean should values containing ocean data be omitted, an integer specifying the cube whose input mask is used to determine land-sea points.

inplace does the function write to an output array inplace or return a single value> defaults to true

ispar boolean to determine if parallelisation should be applied, defaults to true if workers are available.

kwargs additional keyword arguments passed to the inner function

The first argument is always the function to be applied, the second is the input cube or a tuple input cubes if needed. If the function to be applied is registered (either as part of CABLAB or through registerDATFunction), all of the keyword arguments have reasonable defaults and don't need to be supplied. Some of the function still need additional arguments or keyword arguments as is stated in the documentation.

If you want to call mapCube directly on an unregistered function, please have a look at Applying custom functions to get an idea about the usage of the input and output dimensions etc.

source

The function will then be applied to the whole cube in a memory-efficient way, which means that chunks of data are read, processed and then saved in the output cube. Whether the output cube is a TempCube or a CubeMem is decided by the system, depending on if the calculation is parallel, and how large the output cube is.

Here follows a list of analysis function included in this package. If you have implemented or wrapped a method that might be of interest to a broader community, please feel free to open a pull request.

Seasonal cycles

CABLAB.Proc.MSC.gapFillMSC — Method.

gapFillMSC

Fills missing values of each time series with the mean annual cycle.

Call signature

mapCube(gapFillMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis

source

CABLAB.Proc.MSC.getMSC — Function.

getMSC

Returns the mean annual cycle from each time series.

Call signature

mapCube(getMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes MSCaxis

source

CABLAB.Proc.MSC.getMedSC — Method.

getMedMSC

Returns the median annual cycle from each time series.

Call signature

mapCube(getMedMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes MSCaxis

source

CABLAB.Proc.MSC.removeMSC — Method.

removeMSC

Removes the mean annual cycle from each time series.

Call signature

mapCube(removeMSC, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis

source

Outlier detection

CABLAB.Proc.Outlier.DAT_detectAnomalies! — Method.

DAT_detectAnomalies!

A simple wrapper around the function detectAnomalies! from the MultivariateAnomalies package.

Call signature

mapCube(DAT_detectAnomalies!, cube, methods, trainArray)

cube data cube with a axes: TimeAxis, VariableAxis

methods vector of methods to be applied, choose from: KDE,T2,REC,KNN-Gamma,KNN-Delta,SVDD,KNFST

trainArray a matrix of nsample x nvar, to estimate the training parameters for the model. Ideally does not contain any extreme values

Input Axes TimeAxis, Variableaxis

Output Axes TimeAxis, Methodaxis

For an example on how to apply this function, see this notebook.

source

Simple Statistics

Another typcial use case is the application of basic statistics like sum, mean and std. We provide a convenience function reduceCube

CABLAB.DAT.reduceCube — Function.

reduceCube(f::Function, cube, dim::Type{T<:CubeAxis};kwargs...)

Apply a reduction function f on slices of the cube cube. The dimension(s) are specified through dim, which is either an Axis type or a tuple of axis types. Keyword arguments are passed to mapCube or, if unknown passed again to f. It is assumed that f takes an array input and returns a single value.

source

Applying these functions makes sense if the slices one wants to reduce fit in memory. However, if one wants to calculate some statistics on e.g. a timelonlat cube, one would preferably call one of the OnlineStats methods.

Additional simple statistics functions are:

CABLAB.Proc.Stats.normalizeTS — Method.

normalizeTS

Normalize a time series to zeros mean and unit variance

Call signature

mapCube(normalizeTS, cube)

cube data cube with a axes: TimeAxis

Input Axes TimeAxis

Output Axes TimeAxis

source

CABLAB.Proc.Stats.timespacequantiles — Method.

timespacequantiles

Calculate quantiles from a space time data cube. This is usually called on a subset of data returned by sampleLandPoints.

Call signature

mapCube(timespacequantiles, cube, quantiles)

cube data cube with a axes: TimeAxis, SpatialPointAxis

quantiles a vector of quantile values to calculate

Input Axes TimeAxis, SpatialPointAxis

Output Axes QuantileAxis

Calculating exact quantiles from data that don't fit into memory is quite a problem. One solution we provide here is to simply subsample your data and then get the quantiles from a smaller dataset.

For an example on how to apply this function, see this notebook.

source

Time series decomposition

CABLAB.Proc.TSDecomposition.filterTSFFT — Method.

filterTSFFT

Filter each time series using a Fourier filter and return the decomposed series in 4 time windows (Trend, Long-Term Variability, Annual Cycle, Fast Oscillations)

Call signature

mapCube(filterTSFFT, cube)

cube data cube with a axes: TimeAxis

Input Axes Timeaxis

Output Axes Timeaxis, TimeScaleaxis

source

Cube transformations

CABLAB.Proc.CubeIO.extractLonLats — Method.

extractLonLats(c::AbstractCubeData,pl::Matrix)

Extracts a list of longitude/latitude coordinates from a data cube. The coordinates are specified through the matrix pl where size(pl)==(N,2) and N is the number of extracted coordinates. Returns a data cube without LonAxis and LatAxis but with a SpatialPointAxis containing the input locations.

source

CABLAB.Proc.CubeIO.sampleLandPoints — Function.

sampleLandPoints(cube, nsample;nomissing=false)

Get an area-weighted sample from all non-ocean grid cells. This will return a new Cube where the LonAxis and LatAxis are condensed into a single SpatialPointAxis of length nsample. If nomissing=true only grid cells will be selected which don't contain any missing values. This makes sense for gap-filled cubes to make sure that grid cells with systematic seasonal gaps are not selected in the sample.

source

OnlineStats

It is possible to directly apply statistics included in the OnlineStats.jl package on the data cube. This makes it possible to calculate statistics on data too big to fit into memory. The general syntax is

mapCube(f ,cube; by=CubeAxis[], cfun=identity, outAxis=nothing,kwargs...)

where f is an OnlineStat data type and cube is the cube you want to apply the statistics to. By default this function will reduce all values over all axes of the cube, so if you want to do statistics by a certain axis, it has to be specified using the by keyword argument. by accepts a vector of axes types and up to one datacube that can serve as a mask. If such a data cube is supplied, the statistics are split by the unique values in the mask. One can pass a function cfun that transforms the mask values into an index in the range 1..N that defines the index where the new value is going to be put to. If a mask is supplied, it must have either a labels property, which is a Dict{T,String} mapping the numerical mask value to the value name. Alternatively on can supply an outAxis argument that describes the resulting output axis.

This all gets clearer with two small examples. suppose we want to calculate the mean of GPP, NEE and TER under the condition that Tair<280K and Tair>280K over all time steps and grid cells. This is achieved through the following lines of code:

import OnlineStats
lons  = (30,31)
lats  = (50,51)
vars  = ["gross_primary_productivity","net_ecosystem_exchange","terrestrial_ecosystem_respiration"]
t     = getCubeData(ds,variable="air_temperature_2m",longitude=lons,latitude=lats)
cube  = getCubeData(ds,variable=vars,longitude=lons,latitude=lats)

splitTemp(t) = ifelse(t>280,2,1)                            # Define the classification function
outAxis      = CategoricalAxis("TempClass",["< 7C",">7C"])  # A two-length output axis, because there are two possible values
mT    = mapCube(OnlineStats.Mean,cube,by=[t,VariableAxis], cfun=splitTemp, outAxis=outAxis) # Of course we want to split by variable, too

using CABLABPlots
plotXY(mT,xaxis="var",group="tempclass")

A second example would be that we want to calculate averages of the fluxes according to a country mask.

import OnlineStats
vars  = ["gross_primary_productivity","net_ecosystem_exchange","terrestrial_ecosystem_respiration"]
m     = getCubeData(ds,variable="country_mask",longitude=lons,latitude=lats)
cube  = getCubeData(ds,variable=vars,longitude=lons,latitude=lats)

mT    = mapCube(OnlineStats.Mean,cube,by=[m,VariableAxis], cfun=splitTemp, outAxis=outAxis)

This will split the cube by country and variable and compute averages over the input variables.

Online PCA

It is possible to compute a principal component analysis based on a covariance matrix obtained through an online algorithm. The package provides a convenient way to achieve this with the cubePCA function.

CABLAB.Proc.DATOnlineStats.cubePCA — Function.

cubePCA(cube::AbstractCubeData)

Performs a PCA based on a covariance matrix which is estimated through an online algorithm. Returns an OnlinePCA object from which explained_variance and the rotation can be extracted, or which can be used to perform the projection on a dataset.

Keyword arguments

MDAxis specifies the axes that is reduced through the PCA

by a vector of axes types or masks denoting if several PCAs should be performed. If provided, several PCAs will be performed.

noutdims number of output dimensions, how many PCs are estimated

source

For example, if one wants to calculate a PCA over the time dimension, you could use the following code: