Analysis

Apply functions

List of functions

Seasonal cycles

CABLAB.Proc.MSC.getMSC — Function.

Calculate the mean seasonal cycle of xin and write the output to xout.

source

CABLAB.Proc.MSC.getMedSC — Method.

Calculate the median seasonal cycle of xin and write the output to xout.

source

Outlier detection

Simple Statistics

Time series decomposition

Cube transformations