Accessing the Data Cube

Open a datacube

Before one can read data from a cube, it has to be opened. To open a data cube which is accesible through the file system, use the Cube constructor:

using CABLAB
c = Cube("/patch/to/cube")

CABLAB data cube at /Net/Groups/BGI/scratch/DataCube/v0.2.0/high-res/
Spatial resolution:  4320x2160 at 0.083333 degrees.
Temporal resolution: 2001-01-01T00:00:00 to 2012-01-01T00:00:00 at 8daily time steps
Variables:           aerosol_optical_thickness_1610 aerosol_optical_thickness_550 aerosol_optical_thickness_555 aerosol_optical_thickness_659 aerosol_optical_thickness_865 air_temperature_2m bare_soil_evaporation black_sky_albedo burnt_area c_emissions country_mask evaporation evaporative_stress fractional_snow_cover gross_primary_productivity interception_loss land_surface_temperature latent_energy net_ecosystem_exchange open_water_evaporation ozone potential_evaporation precipitation root_moisture sensible_heat snow_sublimation snow_water_equivalent soil_moisture surface_moisture terrestrial_ecosystem_respiration transpiration water_mask water_vapour white_sky_albedo

This returns a Cube object that contains some basics information about the cube which is printed on the screen.

Getting a data handle

var=["c_emissions","air_temperature_2m"]
time=(DateTime("2001-01-01"),DateTime("2001-12-31"))
cubedata = getCubeData(c,longitude=(30,31),latitude=(50,51),variable=var)

Data Cube view with the following dimensions
Lon                 Axis with 4 Elements from 30.0 to 30.75
Lat                 Axis with 4 Elements from 51.0 to 50.25
Time                Axis with 506 Elements from 2001-01-01T00:00:00 to 2011-12-27T00:00:00
Variable            Axis with elements: c_emissions air_temperature_2m

This returns a view into the Data Cube, on which further calculations can be applied. All keyword arguments default to the full range, so calling getCubeData without keyword arguments will return a view No data is read yet. Here you can start to do some calculations on your sub-cube, see either Analysis for a list of methods provided by this framework or Applying custom functions to apply your own functions on the cube. If you just want to visualize the cube see this section Plotting.

Cube Types

In CABLAB, you will

CABLAB.Cubes.AbstractCubeData — Type.

AbstractCubeData{T,N}

Supertype of all cubes. T is the data type of the cube and N the number of dimensions. Beware that an AbstractCubeData does not implement the AbstractArray interface. However, the CABLAB functions mapCube, reduceCube, readCubeData, plotMAP and plotXY will work on any subtype of AbstractCubeData

source

CABLAB.Cubes.CubeMem — Type.

CubeMem{T,N} <: AbstractCubeMem{T,N}

An in-memory data cube. It is returned by applying mapCube when the output cube is small enough to fit in memory or by explicitly calling readCubeData on any type of cube.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

mask N-D array containgin the mask

source

CABLAB.CubeAPI.SubCube — Type.

immutable SubCube{T,C} <: AbstractCubeData{T,4}

A view into the data cube of a single variable. Is the type returned by the mapCube function.

Fields

cube::C Parent cube

variable selected variable

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis

latAxis

timeAxis

source

CABLAB.CubeAPI.SubCubeV — Type.

immutable SubCubeV{T, C} <: AbstractCubeData{T,4}

A view into the data cube with multiple variables. Returned by the mapCube function.

Fields

cube::C Parent cube

variable list of selected variables

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis

latAxis

timeAxis

varAxis

source

CABLAB.Cubes.TempCubes.TempCube — Type.

type TempCube{T,N} <: AbstractCubeData{T,N}

The main data structure for storing temporary results from cube operations. Is usually returned by mapCube, if the result is larger than max_cache

Fields

axes a vector of CubeAxis containing the axes

folder folder containing the data

block_size dimension of the files that the cube is split into

Each TempCube is stored in a single folder, but can contain several files. The rule is that one file is small enough to be read into memory and the block_size determines the size and shape of each sub-file. This data structure is quite convenient for parrallel access, because different processes can write their results at the same time.

source

Cube Axes

Axes are an essential part of each Cube in CABLAB. Every dimension that a cube has is associated with an axis that stores the values of the dimension. For example, a LatitudeAxis will contains a field values representing the chosen latitudes. Similarly, a VariableAxis will contain a list of Variable names. Axes types are divided in categorical axes and axes represented by ranges. All of them are subtypes of the abstract type CubeAxis.

CABLAB.Cubes.Axes.CubeAxis — Type.

abstract CubeAxis{T} <: AbstractCubeData{T,1}

Supertype of all axes. Every CubeAxis is 1D Cube itself and can be passed to mapCube operationes. Although all cube axes are instances of the parametric typealias CategoricalAxis and RangeAxis, there are some typealiases defined to provide shorter and more convenient names for commonly used cube axes. Here is a list of the aliases:

Categorical Axes

VariableAxis represents different variables

SpatialPointAxis represents a list of coordinates

CountryAxis countries

TimeScaleAxis time scales after time series decomposition

QuantileAxis represents different quantiles

Cotinuous Axes

LonAxis longitudes

LatAxis latitudes

TimeAxis time

MSCAxis time step inside a year (for seasonal statistics)

source

CABLAB.Cubes.Axes.CategoricalAxis — Type.

CategoricalAxis{T,S}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol). The default constructor is:

CategoricalAxis(axname::String,values::Vector{T})

source

CABLAB.Cubes.Axes.RangeAxis — Type.

RangeAxis{T,S,R}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol) and R the type of the range which is used to represent the axis values. The default constructor is:

RangeAxis(axname::String,values::Range{T})

source

Cube Masks

Every data cube type in CABLAB contains has a representation for the mask, which has the primary purpose of describing missing values and the reason for missingness. CABLAB masks are represented as UInt8-arrays, where each value can be one of the following:

VALID a regular data entry

MISSING classical missing value

OCEAN masked out by the land-sea mask

OUTOFPERIOD current time step is not inside the measurement period

FILLED does not count as missing, but still denotes that the value is gap filled and not measured

These names can be imported by using CABLAB.Mask. The user can decide if he wants to use the masks in his analyses or rather wants to refer to a different representation with NullableArrays or just representing missings with NaNs. See registerDATFunction for details.

Opening Remote Data Cubes

If you just want to try the CABLAB data cube and don't have access to the full data set, you can open a remote cube through a THREDDS server. All you need is a working internet connection to do this:

CABLAB.CubeAPI.RemoteCube — Type.

Represents a remote data cube accessible through THREDDS. The default constructor is

RemoteCube(base_url)

where base_url is the datacube's base url.

Fields

base_url the cube parent directory

var_name_to_var_index basically the inverse of dataset_files

dataset_files a list of datasets in the cube

dataset_paths a list of urls pointing to the different data sets

config the cube's static configuration CubeConfig

using CABLAB
ds=remoteCube()

source

This will open the remote cube and calling getCubeData will return a cube view that you can process.

Important In order to avoid unnecessary traffic, be nice to our servers. Please use this only for testing the cube software for very limited amount of data (reading maps at single time steps) or time series in lon-lat boxes of size 1degx1deg.

Point-wise access

CABLAB.Proc.CubeIO.sampleLandPoints — Function.

sampleLandPoints(cube, nsample;nomissing=false)

Get an area-weighted sample from all non-ocean grid cells. This will return a new Cube where the LonAxis and LatAxis are condensed into a single SpatialPointAxis of length nsample. If nomissing=true only grid cells will be selected which don't contain any missing values. This makes sense for gap-filled cubes to make sure that grid cells with systematic seasonal gaps are not selected in the sample.

source