Accessing the Data Cube

Accessing the Data Cube

Open a datacube

Before one can read data from a cube, it has to be opened. To open a data cube which is accesible through the file system, use the Cube constructor:

using CABLAB
c = Cube("/patch/to/cube")

CABLAB data cube at /Net/Groups/BGI/scratch/DataCube/v0.2.0/high-res/
Spatial resolution:  4320x2160 at 0.083333 degrees.
Temporal resolution: 2001-01-01T00:00:00 to 2012-01-01T00:00:00 at 8daily time steps
Variables:           aerosol_optical_thickness_1610 aerosol_optical_thickness_550 aerosol_optical_thickness_555 aerosol_optical_thickness_659 aerosol_optical_thickness_865 air_temperature_2m bare_soil_evaporation black_sky_albedo burnt_area c_emissions country_mask evaporation evaporative_stress fractional_snow_cover gross_primary_productivity interception_loss land_surface_temperature latent_energy net_ecosystem_exchange open_water_evaporation ozone potential_evaporation precipitation root_moisture sensible_heat snow_sublimation snow_water_equivalent soil_moisture surface_moisture terrestrial_ecosystem_respiration transpiration water_mask water_vapour white_sky_albedo

This returns a Cube object that contains some basics information about the cube which is printed on the screen.

Getting a data handle

var=["c_emissions","air_temperature_2m"]
time=(Date("2001-01-01"),Date("2001-12-31"))
cubedata = getCubeData(c,longitude=(30,31),latitude=(50,51),time=time,variable=var)

Data Cube view with the following dimensions
Lon                 Axis with 4 Elements from 30.0 to 30.75
Lat                 Axis with 4 Elements from 51.0 to 50.25
Time                Axis with 46 Elements from 2001-01-01 to 2001-12-27
Variable            Axis with elements: c_emissions air_temperature_2m
Total size: 7.19 KB

This returns a view into the Data Cube, on which further calculations can be applied. All keyword arguments default to the full range, so calling getCubeData without keyword arguments will return a view into the whole data cube.

CABLAB.CubeAPI.getCubeData — Function.

getCubeData(cube::Cube;variable,time,latitude,longitude)

Returns a view into the data cube. The following keyword arguments are accepted:

variable: an variable index or name or an iterable returning multiple of these (var1, var2, ...)

time: a single Date object or a 2-element iterable (time_start, time_end)

latitude: a single latitude value or a 2-element iterable (latitude_start, latitude_end)

longitude: a single longitude value or a 2-element iterable (longitude_start, longitude_end)

region: specify a country or SREX region by name or ISO_A3 code. Type ?CABLAB.known_regions to see a list of pre-defined areas

Returns a SubCube object which represents a view into the original data cube.

source

No data is read yet. Here you can start to do some calculations on your sub-cube, see either Analysis for a list of methods provided by this framework or Applying custom functions to apply your own functions on the cube. If you just want to visualize the cube see this section Plotting.

Extracting a list of lon/lat coordinates from a data cube

There are situations in which only a certain list of longitude/latitude pairs is needed for the analysis. One can extract such a list by first creating a cube view containing all the needed variables and then apply the extractLonLats function.

CABLAB.Proc.CubeIO.extractLonLats — Function.

extractLonLats(c::AbstractCubeData,pl::Matrix)

Extracts a list of longitude/latitude coordinates from a data cube. The coordinates are specified through the matrix pl where size(pl)==(N,2) and N is the number of extracted coordinates. Returns a data cube without LonAxis and LatAxis but with a SpatialPointAxis containing the input locations.

source

Here is an example how to apply the function:

cubedata = getCubeData(c,longitude=(30,31),latitude=(50,51),time=time,variable=var)
ll       = [30.1 50.2;
            30.5 51.1;
            30.7 51.1] #Lon/Lats to be extracted
cubenew  = extractLonLats(cubedata,ll)

In-Memory data cube with the following dimensions
Spatial points axis with 3 points
Time                Axis with 46 Elements from 2001-01-01 to 2001-12-27
Variable            Axis with elements: c_emissions air_temperature_2m
Total size: 1.35 KB

Cube Types

In CABLAB, you will

CABLAB.Cubes.AbstractCubeData — Type.

AbstractCubeData{T,N}

Supertype of all cubes. T is the data type of the cube and N the number of dimensions. Beware that an AbstractCubeData does not implement the AbstractArray interface. However, the CABLAB functions mapCube, reduceCube, readCubeData, plotMAP and plotXY will work on any subtype of AbstractCubeData

source

CABLAB.Cubes.CubeMem — Type.

CubeMem{T,N} <: AbstractCubeMem{T,N}

An in-memory data cube. It is returned by applying mapCube when the output cube is small enough to fit in memory or by explicitly calling readCubeData on any type of cube.

Fields

axes a Vector{CubeAxis} containing the Axes of the Cube

data N-D array containing the data

mask N-D array containgin the mask

source

CABLAB.CubeAPI.SubCube — Type.

immutable SubCube{T,C} <: AbstractCubeData{T,4}

A view into the data cube of a single variable. Is the type returned by the mapCube function.

Fields

cube::C Parent cube

variable selected variable

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis

latAxis

timeAxis

source

CABLAB.CubeAPI.SubCubeV — Type.

immutable SubCubeV{T, C} <: AbstractCubeData{T,4}

A view into the data cube with multiple variables. Returned by the mapCube function.

Fields

cube::C Parent cube

variable list of selected variables

sub_grid representation of the subgrid indices

sub_times representation of the selected time steps

lonAxis

latAxis

timeAxis

varAxis

source

CABLAB.Cubes.TempCubes.TempCube — Type.

type TempCube{T,N} <: AbstractCubeData{T,N}

The main data structure for storing temporary results from cube operations. Is usually returned by mapCube, if the result is larger than max_cache

Fields

axes a vector of CubeAxis containing the axes

folder folder containing the data

block_size dimension of the files that the cube is split into

Each TempCube is stored in a single folder, but can contain several files. The rule is that one file is small enough to be read into memory and the block_size determines the size and shape of each sub-file. This data structure is quite convenient for parrallel access, because different processes can write their results at the same time.

source

Cube Axes

Axes are an essential part of each Cube in CABLAB. Every dimension that a cube has is associated with an axis that stores the values of the dimension. For example, a LatitudeAxis will contains a field values representing the chosen latitudes. Similarly, a VariableAxis will contain a list of Variable names. Axes types are divided in categorical axes and axes represented by ranges. All of them are subtypes of the abstract type CubeAxis.

CABLAB.Cubes.Axes.CubeAxis — Type.

abstract CubeAxis{T} <: AbstractCubeData{T,1}

Supertype of all axes. Every CubeAxis is 1D Cube itself and can be passed to mapCube operationes. Although all cube axes are instances of the parametric typealias CategoricalAxis and RangeAxis, there are some typealiases defined to provide shorter and more convenient names for commonly used cube axes. Here is a list of the aliases:

Categorical Axes

VariableAxis represents different variables

SpatialPointAxis represents a list of coordinates

CountryAxis countries

TimeScaleAxis time scales after time series decomposition

QuantileAxis represents different quantiles

Cotinuous Axes

LonAxis longitudes

LatAxis latitudes

TimeAxis time

MSCAxis time step inside a year (for seasonal statistics)

source

CABLAB.Cubes.Axes.CategoricalAxis — Type.

CategoricalAxis{T,S}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol). The default constructor is:

CategoricalAxis(axname::String,values::Vector{T})

source

CABLAB.Cubes.Axes.RangeAxis — Type.

RangeAxis{T,S,R}

To represent axes that are categorical, where T is the element type. The type parameter S denotes the axis name (a symbol) and R the type of the range which is used to represent the axis values. The default constructor is:

RangeAxis(axname::String,values::Range{T})

source

Cube Masks

Every data cube type in CABLAB contains has a representation for the mask, which has the primary purpose of describing missing values and the reason for missingness. CABLAB masks are represented as UInt8-arrays, where each value can be one of the following:

VALID a regular data entry

MISSING classical missing value

OCEAN masked out by the land-sea mask

OUTOFPERIOD current time step is not inside the measurement period

FILLED does not count as missing, but still denotes that the value is gap filled and not measured

These names can be imported by using CABLAB.Mask. The user can decide if he wants to use the masks in his analyses or rather wants to refer to a different representation with NullableArrays or just representing missings with NaNs. See registerDATFunction for details.

Opening Remote Data Cubes

If you just want to try the CABLAB data cube and don't have access to the full data set, you can open a remote cube through a THREDDS server. All you need is a working internet connection to do this:

CABLAB.CubeAPI.RemoteCube — Type.

Represents a remote data cube accessible through THREDDS. The default constructor is

RemoteCube(base_url)

where base_url is the datacube's base url.

Fields

base_url the cube parent directory

var_name_to_var_index basically the inverse of dataset_files

dataset_files a list of datasets in the cube

dataset_paths a list of urls pointing to the different data sets

config the cube's static configuration CubeConfig

using CABLAB
ds=remoteCube()

source

This will open the remote cube and calling getCubeData will return a cube view that you can process.

Important In order to avoid unnecessary traffic, be nice to our servers. Please use this only for testing the cube software for very limited amount of data (reading maps at single time steps) or time series in lon-lat boxes of size 1degx1deg.

Point-wise access

CABLAB.Proc.CubeIO.sampleLandPoints — Function.

sampleLandPoints(cube, nsample;nomissing=false)

Get an area-weighted sample from all non-ocean grid cells. This will return a new Cube where the LonAxis and LatAxis are condensed into a single SpatialPointAxis of length nsample. If nomissing=true only grid cells will be selected which don't contain any missing values. This makes sense for gap-filled cubes to make sure that grid cells with systematic seasonal gaps are not selected in the sample.

source

List of known regions

CABLAB.CubeAPI.known_regions — Constant.

List of Continents

Africa

Asia

Australia

Europe

North America

South America

List of SREX regions

Short Name	Long Name

ALA	Alaska/N.W. Canada

AMZ	Amazon

CAM	Central America/Mexico

CAR	small islands regions Caribbean

CAS	Central Asia

CEU	Central Europe

CGI	Canada/Greenland/Iceland

CNA	Central North America

EAF	East Africa

EAS	East Asia

ENA	East North America

MED	South Europe/Mediterranean

NAS	North Asia

NAU	North Australia

NEB	North-East Brazil

NEU	North Europe

SAF	Southern Africa

SAH	Sahara

SAS	South Asia

SAU	South Australia/New Zealand

SEA	Southeast Asia

SSA	Southeastern South America

TIB	Tibetan Plateau

WAF	West Africa

WAS	West Asia

WNA	West North America

WSA	West Coast South America

ANT	Antarctica

ARC	Arctic

NTP	Pacific Islands region

STP	Southern Topical Pacific

ETP	Pacific Islands region

WIO	West Indian Ocean

List of countries

Short Name	Long Name

ABW	Aruba

AFG	Afghanistan

AGO	Angola

AIA	Anguilla

ALB	Albania

ALA	Aland

AND	Andorra

ARE	United Arab Emirates

ARG	Argentina

ARM	Armenia

ASM	American Samoa

ATA	Antarctica

Ashm	Ashmore and Cartier Is.

ATF	Fr. S. and Antarctic Lands

ATG	Antigua and Barb.

AUS	Australia

AUT	Austria

AZE	Azerbaijan

BDI	Burundi

BEL	Belgium

BEN	Benin

BFA	Burkina Faso

BGD	Bangladesh

BGR	Bulgaria

BHR	Bahrain

BHS	Bahamas

BIH	Bosnia and Herz.

BLM	St. Barthelemy

BLR	Belarus

BLZ	Belize

BMU	Bermuda

BOL	Bolivia

BRA	Brazil

BRB	Barbados

BRN	Brunei

BTN	Bhutan

BWA	Botswana

CAF	Central African Rep.

CAN	Canada

CHE	Switzerland

CHL	Chile

CHN	China

CIV	Ivory Coast

CLP	Clipperton I.

CMR	Cameroon

CNM

COD	Congo (Kinshasa)

COG	Congo (Brazzaville)

COK	Cook Is.

COL	Colombia

COM	Comoros

CPV	Cape Verde

CRI	Costa Rica

CSI	Coral Sea Is.

CUB	Cuba

CUW	Curacao

CYM	Cayman Is.

CYN	N. Cyprus

CYP	Cyprus

CZE	Czech Rep.

DEU	Germany

DJI	Djibouti

DMA	Dominica

DNK	Denmark

DOM	Dominican Rep.

DZA	Algeria

ECU	Ecuador

EGY	Egypt

ERI	Eritrea

ESB	Dhekelia

ESP	Spain

EST	Estonia

ETH	Ethiopia

FIN	Finland

FJI	Fiji

FLK	Falkland Is.

FRA	France

FRO	Faroe Is.

FSM	Micronesia

GAB	Gabon

Gaza	Gaza

GBR	United Kingdom

GEO	Georgia

GGY	Guernsey

GHA	Ghana

GIB	Gibraltar

GIN	Guinea

GMB	Gambia

GNB	Guinea Bissau

GNQ	Eq. Guinea

GRC	Greece

GRD	Grenada

GRL	Greenland

GTM	Guatemala

GUM	Guam

GUY	Guyana

HKG	Hong Kong

HMD	Heard I. and McDonald Is.

HND	Honduras

HRV	Croatia

HTI	Haiti

HUN	Hungary

IDN	Indonesia

IMN	Isle of Man

IND	India

IOA	Indian Ocean Ter.

IOT	Br. Indian Ocean Ter.

IRL	Ireland

IRN	Iran

IRQ	Iraq

ISL	Iceland

ISR	Israel

ITA	Italy

JAM	Jamaica

JEY	Jersey

JOR	Jordan

JPN	Japan

KAB	Baykonur

KAS

KAZ	Kazakhstan

KEN	Kenya

KGZ	Kyrgyzstan

KHM	Cambodia

KIR	Kiribati

KNA	St. Kitts and Nevis

KNM

KOR	S. Korea

KOS	Kosovo

KWT	Kuwait

LAO	Laos

LBN	Lebanon

LBR	Liberia

LBY	Libya

LCA	Saint Lucia

LIE	Liechtenstein

LKA	Sri Lanka

LSO	Lesotho

LTU	Lithuania

LUX	Luxembourg

LVA	Latvia

MAC	Macau

MAF	St. Martin

MAR	Morocco

MCO	Monaco

MDA	Moldova

MDG	Madagascar

MDV	Maldives

MEX	Mexico

MHL	Marshall Is.

MKD	Macedonia

MLI	Mali

MLT	Malta

MMR	Myanmar

MNE	Montenegro

MNG	Mongolia

MNP	N. Mariana Is.

MOZ	Mozambique

MRT	Mauritania

MSR	Montserrat

MUS	Mauritius

MWI	Malawi

MYS	Malaysia

NAM	Namibia

NCL	New Caledonia

NER	Niger

NFK	Norfolk Island

NGA	Nigeria

NIC	Nicaragua

NIU	Niue

NLD	Netherlands

NOR	Norway

NPL	Nepal

NRU	Nauru

NZL	New Zealand

OMN	Oman

PAK	Pakistan

PAN	Panama

PCN	Pitcairn Is.

PER	Peru

PHL	Philippines

PLW	Palau

PNG	Papua New Guinea

POL	Poland

PRI	Puerto Rico

PRK	N. Korea

PRT	Portugal

PRY	Paraguay

PYF	Fr. Polynesia

QAT	Qatar

ROU	Romania

RUS	Russia

RWA	Rwanda

SAH	W. Sahara

SAU	Saudi Arabia

SDN	Sudan

SDS	S. Sudan

SEN	Senegal

SGP	Singapore

SGS	S. Geo. and S. Sandw. Is.

SHN	Saint Helena

SLB	Solomon Is.

SLE	Sierra Leone

SLV	El Salvador

SMR	San Marino

SOL	Somaliland

SOM	Somalia

SPM	St. Pierre and Miquelon

SRB	Serbia

STP	Sao Tome and Principe

SUR	Suriname

SVK	Slovakia

SVN	Slovenia

SWE	Sweden

SWZ	Swaziland

SXM	Sint Maarten

SYC	Seychelles

SYR	Syria

TCA	Turks and Caicos Is.

TCD	Chad

TGO	Togo

THA	Thailand

TJK	Tajikistan

TKM	Turkmenistan

TLS	East Timor

TON	Tonga

TTO	Trinidad and Tobago

TUN	Tunisia

TUR	Turkey

TUV	Tuvalu

TWN	Taiwan

TZA	Tanzania

UGA	Uganda

UKR	Ukraine

UMI	U.S. Minor Outlying Is.

URY	Uruguay

USA	United States

USG	Guantanamo Bay USNB

UZB	Uzbekistan

VAT	Vatican

VCT	St. Vin. and Gren.

VEN	Venezuela

VGB	British Virgin Is.

VIR	U.S. Virgin Is.

VNM	Vietnam

VUT	Vanuatu

PSE	West Bank

WLF	Wallis and Futuna

WSB	Akrotiri

WSM	Samoa

YEM	Yemen

ZAF	South Africa

ZMB	Zambia

ZWE	Zimbabwe

source