Loading and saving results

ESDL.Cubes.saveCube — Function

saveCube(cube,name::String)

Save a ZarrCube or CubeMem to the folder name in the ESDL working directory.

See also loadCube

ESDL.Cubes.ESDLZarr.loadCube — Function

loadCube(name::String)

Loads a cube that was previously saved with saveCube. Returns a TempCube object.

ESDL.ESDLTools.@loadOrGenerate — Macro

macro loadOrGenerate(x...,expression)

Takes a list of variablename=>"Storage Name" pairs. Checks if all datasets can be found on disk and loads them. If not, the datasets will be regenerated by evaluating the given expression.

To force recalculation, call ESDL.recalculate(true) before evaluating the macro.

Example

The following lines will check if cubes with the names "Filled" and "Normalized" exist on disk, load them and assign the variable names cube_filled and cube_norm. If the datasets to not exist on disk, they are generated and saved under the given names.

@loadOrGenerate cube_filled=>"Filled" cube_norm=>"Normalized" begin
cube_filled = mapCube(gapFillMSC,d)
cube_norm   = mapCube(normalize_TS,d)
end