Plotting

Plot geographical maps

Map plotting is generally done using the plotMAP function:

CABLAB.Plot.plotMAP — Function.

plotMAP(cube::AbstractCubeData; dmin=datamin, dmax=datamax, colorm=colormap("oranges"), oceancol=colorant"darkblue", misscol=colorant"gray", kwargs...)

Map plotting tool for cube objects, can be called on any type of cube data

Keyword arguments

dmin, dmax Minimum and maximum value to be used for color transformation

colorm colormap to be used. Find a list of colormaps in the Colors.jl package

oceancol color to fill the ocean with, defaults to colorant"darkblue"

misscol color to represent missing values, defaults to colorant"gray"

symmetric make the color scale symmetric around zero

labels given a list of labels this will create a plot with a non-continouous color scale where integer cube values [1..N] are mapped to the given labels.

dim=value can set other dimensions to certain values, for example var="air_temperature_2m" will fix the variable for the resulting plot

If a dimension is neither longitude or latitude and is not fixed through an additional keyword, a slider or dropdown menu will appear to select the axis value.

source

Here is an example on how to plot a map. The keyword arguments specify the time step (time=1) and the variable (var=1).

cdata=getCubeData(ds,variable=["air_temperature_2m","gross_primary_productivity"])
plotMAP(cdata,time=1,var=1)

Inside a Jupyter notebook, the keyword arguments can be omitted and sliders or dropdown menus will be shown to select the desired values.

Other plots

XY plots

Generating x-y type plots where the x axis is one of the cube axes and the y axis is the corresponding cube value is done with the generic plotXY function.

CABLAB.Plot.plotXY — Function.

plotXY(cube::AbstractCubeData; group=0, xaxis=-1, kwargs...)

Generic plotting tool for cube objects, can be called on any type of cube data.

Keyword arguments

xaxis which axis is to be used as x axis. Can be either an axis Datatype or a string. Short versions of axes names are possible as long as the axis can be uniquely determined.

group it is possible to group the plot by a categorical axis. Can be either an axis data type or a string.

dim=value can set other dimensions to certain values, for example lon=51.5 will fix the longitude for the resulting plot

If a dimension is not the x axis or group variable and is not fixed through an additional keyword, a slider or dropdown menu will appear to select the axis value.

source

Here are two examples for using this function:

cdata=getCubeData(ds,variable=["net_ecosystem_exchange","gross_primary_productivity","terrestrial_ecosystem_respiration"],
longitude=(30.0,30.0),latitude=(50.0,52.0))
plotXY(cdata,xaxis="time",group="variable",lon=31,lat=51)
nothing # hide

This is a time series plot, grouped by variables for a specific longitude/latitude.

m=reduceCube(mean,cdata,TimeAxis)
plotXY(m,xaxis="variable",group="lat",lon=30)

Scatter plots

In order to do scatter plots, i.e. plotting variable A against variable B one can use the plotScatter function.

plotScatter

A short example is shown here:

cdata=getCubeData(ds,variable=["net_ecosystem_exchange","gross_primary_productivity","terrestrial_ecosystem_respiration"],
longitude=(30.0,30.0),latitude=(50.0,52.0))
p=plotScatter(cdata,alongaxis=TimeAxis,xaxis=1,yaxis=2,group="lat",lon=30)